requirements.txt
Django>=4.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2